docs: typings for version
